Move Grouping Use Cases Into Namespace
